fix: resolve all clippy and rustfmt errors in backend
- Fix if_same_then_else: collapse redundant model selection branches in ai.rs
- Fix match_unwrap_or_default: simplify gate lookup in sdk.rs
- Add clippy.toml with too-many-arguments-threshold = 10
- Add #![allow(dead_code)] for intentional domain model structs
- Apply cargo fmt to ai_service.rs and polling_service.rs

+++ backend/src/services/tracking_service.rs
@@ -207,7 +207,7 @@ impl TrackingService {
             .map(Self::row_to_session)
             .collect::<Result<_>>()?;
         if let Some(ttl_minutes) = self.session_ttl_minutes {
-            let ttl_seconds = ttl_minutes.saturating_mul(60) as i64;
+            let ttl_seconds = ttl_minutes.saturating_mul(60);
             for session in &mut sessions {
                 if session.ended_at.is_none() {
                     let elapsed = now.signed_duration_since(session.started_at).num_seconds();
